Implemented Field Status Board ROS adapter for Gamefield and Machines

+++ include/elements/FieldArea.h
@@ -18,7 +18,7 @@ namespace rcll_draw {
         void setGeometry(int x, int y, int w, int h, int gapsize);
         void setGameInfo(std::string gamestate, std::string gamephase, int time, int score_cyan, int score_magenta);
         void setLayout(double field_w, double field_h, int zones_x, int zones_y, std::vector<int> insertion_zones);
-        void setWalls(std::vector<double> wall_coordinates);
+        void setWalls(std::vector<float> wall_coordinates);
         void setTeam(std::string team_name, rcll_draw::Team team_color);
         size_t addRobot(std::string name, int id, rcll_draw::Team team);
         void setRobotPos(double x, double y, double yaw, size_t index);
